backend/requirements.txt
# FastAPI и веб-сервер
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Базы данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
neo4j==5.14.1
redis==5.0.1

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# HTTP клиенты
httpx==0.28.1
requests==2.32.5

# Обработка документов
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
docx2pdf==0.1.8
pytesseract==0.3.10
Pillow==12.0.0

# ИИ и ML
llama-cpp-python==0.2.20
sentence-transformers==2.2.2
transformers==4.57.1
huggingface-hub==0.35.2
aiogram==3.4.1
unsloth==2025.11.2
accelerate==1.11.0
bitsandbytes==0.48.2
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
xformers==0.0.32.post2
datasets==4.3.0
trl==0.23.0
peft==0.17.1
numpy==2.3.4
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
lxml==5.3.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
openai==1.54.0
xlrd==2.0.1
mammoth==1.6.0
beautifulsoup4==4.12.3
ezdxf==1.3.8
ifcopenshell==0.7.0

# MinIO
minio==7.2.0

# Очереди задач
celery==5.3.4
flower==2.0.1

# Логирование
structlog==23.2.0

# Утилиты
python-dotenv==1.0.0
pytz==2025.2

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

